# docs/openapi-v1.yaml
openapi: 3.1.0
info:
  title: JARVIS Backend API
  version: 1.1.0
  description: |
    v1 API contract aligned with currently implemented Fastify routes.
servers:
  - url: https://api.jarvis.local
    description: Default server
...
paths:
...
  /api/v1/executions/runs:
    post:
      tags: [Executions]
      summary: Start asynchronous execution run for code/compute workloads
      operationId: startExecutionRun
      parameters:
        - in: header
          name: idempotency-key
          required: true
          schema:
            type: string
            minLength: 8
            maxLength: 200
        - in: header
          name: x-trace-id
          required: false
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExecutionRunCreateRequest'
      responses:
        '202':
          description: Execution run accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiSuccessExecutionRun'
        '200':
          description: Existing execution run returned for idempotency replay
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiSuccessExecutionRun'
        default:
          $ref: '#/components/responses/ApiError'
    get:
      tags: [Executions]
      summary: List recent execution runs
      operationId: listExecutionRuns
      parameters:
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        '200':
          description: Execution run list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiSuccessExecutionRuns'
        default:
          $ref: '#/components/responses/ApiError'
  /api/v1/executions/runs/{runId}:
    get:
      tags: [Executions]
      summary: Get single execution run
      operationId: getExecutionRun
      parameters:
        - in: path
          name: runId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Execution run detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiSuccessExecutionRun'
        '404':
          $ref: '#/components/responses/ApiError'
        default:
          $ref: '#/components/responses/ApiError'
  /api/v1/executions/runs/{runId}/events:
    get:
      tags: [Events]
      summary: Stream execution run lifecycle events over SSE
      operationId: streamExecutionRunEvents
      parameters:
        - in: path
          name: runId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 'text/event-stream payload (event: stream.open|execution.run.updated|execution.run.completed|execution.run.failed|stream.close)'
          content:
            text/event-stream:
              schema:
                $ref: '#/components/schemas/ExecutionRunSseDataEnvelope'
              examples:
                streamOpen:
                  summary: event stream.open
                  value:
                    request_id: req_123
                    run_id: run_123
                runUpdated:
                  summary: event execution.run.updated
                  value:
                    run_id: run_123
                    timestamp: '2026-02-23T08:31:05.100Z'
                    data:
                      id: run_123
                      mode: code
                      prompt: Write test cases for user auth flow.
                      status: running
                      output: Execution run started.
                      attempts: []
                      provider: null
                      model: pending
                      used_fallback: false
                      task_id: null
                      duration_ms: 0
                      created_at: '2026-02-23T08:31:04.000Z'
                      updated_at: '2026-02-23T08:31:05.000Z'
                streamClose:
                  summary: event stream.close
                  value:
                    run_id: run_123
        '404':
          $ref: '#/components/responses/ApiError'
        default:
          $ref: '#/components/responses/ApiError'
  /api/v1/tasks:
    post:
      tags: [Tasks]
      summary: Create task
      operationId: createTask
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskCreateRequest'
      responses:
        '201':
          description: Task created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiSuccessTask'
        default:
          $ref: '#/components/responses/ApiError'

    get:
      tags: [Tasks]
      summary: List tasks
      operationId: listTasks
      parameters:
        - in: query
          name: status
          schema:
            type: string
            enum: [queued, running, blocked, retrying, done, failed, cancelled]
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 200
            default: 50
      responses:
        '200':
          description: Task list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiSuccessTaskList'
        default:
          $ref: '#/components/responses/ApiError'
...
components:
  responses:
    ApiError:
      description: Standard API error envelope
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorEnvelope'
  schemas:
    ApiMeta:
      type: object
      additionalProperties: true
    ApiErrorEnvelope:
      type: object
      required: [request_id, error]
      properties:
        request_id:
          type: string
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: object
              additionalProperties: true
...
    ProviderAttempt:
      type: object
      required: [provider, status]
      properties:
        provider:
          type: string
          enum: [openai, gemini, anthropic, local]
        status:
          type: string
          enum: [success, failed, skipped]
        latencyMs:
          type: integer
        error:
          type: string
...
    ExecutionRunCreateRequest:
      type: object
      required: [mode, prompt]
      properties:
        mode:
          type: string
          enum: [code, compute]
        prompt:
          type: string
          minLength: 1
          maxLength: 8000
        system_prompt:
          type: string
        provider:
          type: string
          enum: [auto, openai, gemini, anthropic, local]
          default: auto
        exclude_providers:
          type: array
          maxItems: 4
          items:
            type: string
            enum: [openai, gemini, anthropic, local]
        strict_provider:
          type: boolean
          default: false
        model:
          type: string
        temperature:
          type: number
          minimum: 0
          maximum: 2
        max_output_tokens:
          type: integer
          minimum: 1
          maximum: 32000
        create_task:
          type: boolean
          default: true
        task_title:
          type: string
          minLength: 1
          maxLength: 200
    ExecutionRun:
      type: object
      required: [id, mode, prompt, status, output, attempts, provider, model, used_fallback, task_id, duration_ms, created_at, updated_at]
      properties:
        id:
          type: string
        mode:
          type: string
          enum: [code, compute]
        prompt:
          type: string
        status:
          type: string
          enum: [queued, running, completed, failed]
        output:
          type: string
        attempts:
          type: array
          items:
            $ref: '#/components/schemas/ProviderAttempt'
        provider:
          type: string
          enum: [openai, gemini, anthropic, local]
          nullable: true
        model:
          type: string
        used_fallback:
          type: boolean
        task_id:
          type: string
          nullable: true
        duration_ms:
          type: integer
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    Task:
      type: object
      required: [id, userId, mode, status, title, input, idempotencyKey, createdAt, updatedAt]
      properties:
        id:
          type: string
        userId:
          type: string
        mode:
          type: string
          enum: [chat, execute, council, code, compute, long_run, high_risk, radar_review, upgrade_execution]
        status:
          type: string
          enum: [queued, running, blocked, retrying, done, failed, cancelled]
        title:
          type: string
        input:
          type: object
          additionalProperties: true
        idempotencyKey:
          type: string
        traceId:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    TaskCreateRequest:
      type: object
      required: [title]
      properties:
        mode:
          type: string
          enum: [chat, execute, council, code, compute, long_run, high_risk, radar_review, upgrade_execution]
          default: execute
        title:
          type: string
          minLength: 1
          maxLength: 200
        input:
          type: object
          additionalProperties: true
...
    ExecutionRunStreamOpen:
      type: object
      required: [request_id, run_id]
      properties:
        request_id:
          type: string
        run_id:
          type: string
    ExecutionRunStreamClose:
      type: object
      required: [run_id]
      properties:
        run_id:
          type: string
    ExecutionRunStreamUpdate:
      type: object
      required: [run_id, timestamp, data]
      properties:
        run_id:
          type: string
        timestamp:
          type: string
          format: date-time
        data:
          $ref: '#/components/schemas/ExecutionRun'
    ExecutionRunSseDataEnvelope:
      oneOf:
        - $ref: '#/components/schemas/ExecutionRunStreamOpen'
        - $ref: '#/components/schemas/ExecutionRunStreamClose'
        - $ref: '#/components/schemas/ExecutionRunStreamUpdate'
...
    ApiSuccessExecutionRun:
      type: object
      required: [request_id, data, meta]
      properties:
        request_id:
          type: string
        data:
          $ref: '#/components/schemas/ExecutionRun'
        meta:
          $ref: '#/components/schemas/ApiMeta'
    ApiSuccessExecutionRuns:
      type: object
      required: [request_id, data, meta]
      properties:
        request_id:
          type: string
        data:
          type: object
          required: [runs]
          properties:
            runs:
              type: array
              items:
                $ref: '#/components/schemas/ExecutionRun'
        meta:
          $ref: '#/components/schemas/ApiMeta'
    ApiSuccessTask:
      type: object
      required: [request_id, data, meta]
      properties:
        request_id:
          type: string
        data:
          $ref: '#/components/schemas/Task'
        meta:
          $ref: '#/components/schemas/ApiMeta'
    ApiSuccessTaskList:
      type: object
      required: [request_id, data, meta]
      properties:
        request_id:
          type: string
        data:
          type: array
          items:
            $ref: '#/components/schemas/Task'
        meta:
          $ref: '#/components/schemas/ApiMeta'
...